Add a PXD only Cython extension for the WebSocket reader to improve performance (#9543)
(cherry picked from commit 6ae2f053e9989a759b140e42937e2138f80e502d)
