+++ CHANGES/9543.feature.rst
@@ -1,0 +1,1 @@
+Improved performance of reading WebSocket messages with a Cython implementation -- by :user:`bdraco`.
