+++ aiohttp/_websocket/mask.pxd
@@ -1,0 +1,3 @@
+"""Cython declarations for websocket masking."""
+
+cpdef void _websocket_mask_cython(bytes mask, bytearray data)
