+++ aiohttp/_websocket/reader_c.py
@@ -1,0 +1,1 @@
+reader_py.py
